+++ packages/EasyDoctrine/src/Interfaces/EntityEventSubscriberInterface.php
@@ -7,6 +7,9 @@ use Doctrine\Common\EventSubscriber;
 use Doctrine\ORM\Event\OnFlushEventArgs;
 use Doctrine\ORM\Event\PostFlushEventArgs;
 
+/**
+ * @deprecated since 5.11, will be removed in 6.0. Use \EonX\EasyDoctrine\Listeners\EntityEventListener instead.
+ */
 interface EntityEventSubscriberInterface extends EventSubscriber
 {
     /**
